+++ ln/docs.go
@@ -1,0 +1,6 @@
+/*
+Package ln contains the Lightning Network-related code.
+
+There should be no need to use this package when implementing a web service that uses ln-paywall.
+*/
+package ln