+++ storage/docs.go
@@ -1,0 +1,7 @@
+/*
+Package storage contains storage implementations and related options.
+
+These storage implementations satisfy the wall.StorageClient interface.
+You can use one of them or implement your own.
+*/
+package storage